backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9 
pydantic==1.9.0
pydantic[email]==1.9.0
python-dotenv==0.19.0
alembic==1.13.1